Add support for displaying an optional tag passed to the tracer
Dynamically named anonymous functions aren't always a feasible option. For example the recommended debugging workflow for re-frame conflicts with the usage of plumatic-schema-annotated functions (nested s/defn macros pased to the trace-forms macro), so this is an easy solution to add some additional information to a trace like p.ex. the handler name.